+++ jest.config.ts
@@ -1,6 +1,7 @@
 module.exports = {
   rootDir: "src",
   preset: "ts-jest",
+  testRegex: "src/.*\\.test\\.ts$",
   testEnvironment: "node",
-  verbose: true,
+  moduleFileExtensions: ["ts", "js"],
 };