+++ src/validators/date/utils.ts
@@ -1,6 +1,7 @@
 export * from "./required";
 export * from "./min";
 export * from "./max";
+export * from "./notIn";
 
 export const DEFAULT_MESSAGE = {
   required: "The value is required!",
@@ -10,6 +11,7 @@ export const DEFAULT_MESSAGE = {
     `Must be less than or equal to ${new Date(value)}!`,
   between: (min: Date | number, max: Date | number) =>
     `Must be between ${new Date(min)} and ${new Date(max)}!`,
+  notIn: "The value is not allowed!",
 };
 
 export type VesoDateValueTypes = "date" | "null" | "undefined";
@@ -17,7 +19,8 @@ export type VesoDateValueTypes = "date" | "null" | "undefined";
 export type VesoDateCheck =
   | { type: "required"; message: string }
   | { type: "min"; value: Date | number; message: string }
-  | { type: "max"; value: Date | number; message: string };
+  | { type: "max"; value: Date | number; message: string }
+  | { type: "notIn"; value: Date[]; message: string };
 
 export type VesoDateConstructor = {
   check?: VesoDateCheck[];